Simplified version constraints accross all cabal files
Versions constraint propagate to dependencies, which means that if
a test component depends on a library in the same cabal file, then it
will always use the same versions as the library.  For this reason
there's no need to over constraint executables, tests and benchmarks
with constrains which are already present in the library component.