stylish-haskell: fix stylish-haskell make target
1. Ensure .stylish-haskell.yaml is picked up properly.
2. Avoid descending into inappropriate subdirectories.
3. Pick up language extensions from the sub-projects' cabal files.

Failures to parse source files occurred by dint of not having language
extensions enabled according to the sub-projects' cabal files and the
global .yaml file. The global .yaml file's path is passed in as an
argument to the -c option at each invocation. It also descends into
sub-projects' directories to launch stylish-haskell invocations in order
to accommodate the limitations on stylish-haskell's procedure to search
for cabal files.

Co-authored-by: Nadia Chambers <nadia.chambers@iohk.io>
Co-authored-by: Clément Hurlin <clement.hurlin@iohk.io>